Splitup sink_cluster.rs
